refactor[#268] : app 레이어 구조 변경에 따른 테스트 코드 변경
