+++ src/pages/login/email/email.stories.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router-dom";
 import { Meta, StoryObj } from "@storybook/react";
-import MainLayout from "@/pages/layout";
+import GlobalLayout from "@/app/GlobalLayout";
 import EmailLoginPage from "./page";
 
 const meta: Meta<typeof EmailLoginPage> = {
@@ -23,7 +23,7 @@ type Story = StoryObj<typeof EmailLoginPage>;
 export const Default: Story = {
   render: () => (
     <Routes>
-      <Route element={<MainLayout />}>
+      <Route element={<GlobalLayout />}>
         <Route index element={<EmailLoginPage />} />
       </Route>
     </Routes>
